导航链接、文件上传、细节优化

+++ portal/src/components/manager/file/Img.vue
@@ -21,27 +21,36 @@
             :total="fileListData.total">
             </el-pagination>
             <br>
+            <!-- 提示信息 -->
+            <el-alert
+                title="温馨提示"
+                type="warning"
+                :closable="false"
+                description="">
+                查看图片时小心不要点到删除，如要删除请确保没有文章引用。不建议进行删除操作。
+            </el-alert>
+            <br>
             <el-upload
-            accept=".jpg,.png,.bmp,.gif"
-            action=""
-            :http-request="uploadFile"
-            list-type="picture-card"
-            :file-list="fileListData.dataArray"
-            :on-preview="handlePictureCardPreview"
-            :on-remove="deleFile">
-            <i class="el-icon-plus"></i>
+              accept=".jpg,.png,.bmp,.gif"
+              action=""
+              :http-request="uploadFile"
+              list-type="picture-card"
+              :file-list="fileListData.dataArray"
+              :on-preview="handlePictureCardPreview"
+              :on-remove="deleFile">
+              <i class="el-icon-plus"></i>
             </el-upload>
             <el-dialog :visible.sync="dialogFile.visible" size="tiny" :title="dialogFile.file.name">
                 <el-alert
                   title="文件直链"
                   type="success"
-                  :closable="false"
-                  :description="dialogFile.file.url">
+                  :closable="false">
+                  <a :href="dialogFile.file.url" style="color:black" target="_blank">{{dialogFile.file.url}}</a>
                 </el-alert>
                 <br>
                 <img width="100%" :src="dialogFile.file.url" :alt="dialogFile.file.description">
             </el-dialog>
-            
+
         </el-card>
     </div>
 </template>
@@ -74,11 +83,24 @@
     methods: {
       // 删除文件
       async deleFile(file, fileList) {
-        this.$message.warning('将要删除文件'+ file.fid)
-        const {data: res} = await this.$http.delete(`file/${file.fid}`)
-        console.log(res)
-        if(res.meta.status !== 200) return this.$message.error('后台接口异常，删除此文件失败！返回信息：'+res.meta.msg)
-        else this.$message.success(res.meta.msg)
+        this.$confirm('此操作将永久删除该文件, 是否继续?', '提示', {
+            confirmButtonText: '确定删除',
+            cancelButtonText: '取消',
+            type: 'warning'
+            }).then(async () => {
+                // 请求后台，删除文件
+                this.$message.warning('将要删除文件'+ file.fid)
+                const {data: res} = await this.$http.delete(`file/${file.fid}`)
+                console.log(res)
+                if(res.meta.status !== 200) return this.$message.error('后台接口异常，删除此文件失败！返回信息：'+res.meta.msg)
+                else this.$message.success(res.meta.msg)
+            }).catch(() => {
+                this.$message({
+                    type: 'success',
+                    message: '已取消删除, 刷新下可看到文件。'
+                });          
+            });
+        
       },
       //   显示图片大图
       handlePictureCardPreview(file) {
@@ -94,11 +116,12 @@
         console.log(res)
         if(res.meta.status !== 200) return this.$message.error('后台接口异常，上传图片失败！返回信息：'+res.meta.msg)
         // 文件上传成功提示信息
-        this.$message.success('文件上传成功')
+        this.$message.success('文件上传成功');
+        this.getFileListData()
       },
       // 获取文件数据
       async getFileListData(){
-        const {data: res} = await this.$http.get(`file/list/img/${this.fileListData.currentPage}/${this.fileListData.pageSize}`)
+        const {data: res} = await this.$http.get(`file/list/${this.fileListData.currentPage}/${this.fileListData.pageSize}`)
         console.log(res)
         if(res.meta.status !== 200) return this.$message.error('后台接口异常，获取所有文件失败！返回信息：'+res.meta.msg)
         else this.$message.success(res.meta.msg)
@@ -118,3 +141,5 @@
   }
 </script>
 
+<style lang="less" scoped>
+</style>
